requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
psycopg2-binary
Werkzeug
gunicorn  # chỉ cần cho Heroku
